requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alpha_vantage==3.0.0
anyio==4.4.0
APScheduler==3.6.3
async-timeout==4.0.3
attrs==24.2.0
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
forex-python==1.8
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
multidict==6.1.0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
tornado==6.4.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
yarl==1.11.1
